Starts perl call code refactore

+++ src/main.cpp
@@ -15,11 +15,11 @@ using namespace std;
 using namespace libconfig;
 using namespace TgBot;
 
-static PerlInterpreter *my_perl;  /***    The Perl interpreter    ***/
+static PerlInterpreter *perl_int;  /***    The Perl interpreter    ***/
 
 bool sigintGot = false;
 
-int getToken(const string &filepath, string &token) {
+int get_token(const string &filepath, string &token) {
     Config keys;
 
     // Read the file. If there is an error, report it and exit.
@@ -41,7 +41,7 @@ int getToken(const string &filepath, string &token) {
     return 0;
 }
 
-void setupBot(Bot &bot) {
+void setup_bot(Bot &bot) {
     bot.getEvents().onCommand("start", [&bot](Message::Ptr message) -> void {
         bot.getApi().sendMessage(message->chat->id, "Bip Bop! Getting the motor running...");
     });
@@ -53,47 +53,63 @@ void setupBot(Bot &bot) {
     bot.getEvents().onNonCommandMessage([&bot](Message::Ptr message) -> void {
         bool from_group = ((message->chat->type == Chat::Type::Group) or (message->chat->type == Chat::Type::Supergroup));
         if (!from_group) {
-            cout << "Private message from: " << message->from->username << endl;
+            cout << "[" << message->from->username << "]\t" << message->text << endl;
+            // ignore for now, maybe block user
             return;
         }
-        cout << "Chat ID: " << message->chat->id << "\tID: " << message->messageId << endl;
-        cout << message->from->username << ":\t" << message->text << endl;
-        bot.getApi().sendMessage(message->chat->id, "Deleting: " + message->text);
-        bot.getApi().deleteMessage(message->chat->id, message->messageId);
+
+        // Log message
+        cout << "[" << message->chat->title << "]" << endl;
+        cout << ">> " << message->from->username << ":\t" << message->text << endl;
+
+        // Debug
+        if(is_spam_message(message->text)) {
+            bot.getApi().sendMessage(message->chat->id, "Deleting: " + message->text);
+            bot.getApi().deleteMessage(message->chat->id, message->messageId);
+        }
     });
+}
+
+void setup_perl(PerlInterpreter *perl_int, const string module_path="hello.pl") {
+    perl_int = perl_alloc();
+    perl_construct(perl_int);
+    PL_exit_flags |= PERL_EXIT_DESTRUCT_END;
+
+    char *perl_argv[] = {"", module_path.c_str()};
+    perl_parse(perl_int, xs_init, 2, perl_argv, (char **)NULL);
+    perl_run(perl_int);
+}
+
+void destroy_perl(PerlInterpreter *perl_int) {
+    perl_destruct(perl_int);
+    perl_free(perl_int);
+    PERL_SYS_TERM();
+}
+
+bool is_spam_message(const string message, const string spam_func="validate") {
+    dSP;                                                                /* initialize stack pointer      */
+    ENTER;                                                              /* everything created after here */
+    SAVETMPS;                                                           /* ...is a temporary variable.   */
+    PUSHMARK(SP);                                                       /* remember the stack pointer    */
+    XPUSHs(sv_2mortal(newSVpv(message.c_str(), message.length())));     /* push the base onto the stack  */
+    PUTBACK;                                                            /* make local stack pointer global */
+    call_pv(spam_func.c_str(), G_SCALAR);                               /* call the function             */
+    SPAGAIN;                                                            /* refresh stack pointer         */
+    bool is_spam = (bool) POPi;                                         /* pop the return value from stack */
+    PUTBACK;
+    FREETMPS;                                                           /* free that return value        */
+    LEAVE;                                                              /* ...and the XPUSHed "mortal" args.*/
+    return is_spam;
 }
 
 int main(int argc, char **argv, char **env) {
 
-    char *perl_argv[] = {"", "hello.pl"};
     PERL_SYS_INIT3(&argc,&argv,&env);
-    my_perl = perl_alloc();
-    perl_construct(my_perl);
-    PL_exit_flags |= PERL_EXIT_DESTRUCT_END;
-    perl_parse(my_perl, xs_init, 2, perl_argv, (char **)NULL);
-    perl_run(my_perl);
 
-    dSP;                                        /* initialize stack pointer      */
-    ENTER;                                      /* everything created after here */
-    SAVETMPS;                                   /* ...is a temporary variable.   */
-    PUSHMARK(SP);                               /* remember the stack pointer    */
-    XPUSHs(sv_2mortal(newSVpv("test", 4)));     /* push the base onto the stack  */
-    PUTBACK;                                    /* make local stack pointer global */
-    call_pv("validate", G_SCALAR);              /* call the function             */
-    SPAGAIN;                                    /* refresh stack pointer         */
-    int val = POPi;                                    /* pop the return value from stack */
-    PUTBACK;
-    FREETMPS;                                   /* free that return value        */
-    LEAVE;                                      /* ...and the XPUSHed "mortal" args.*/
-
-    perl_destruct(my_perl);
-    perl_free(my_perl);
-    PERL_SYS_TERM();
-
-    std::cout << "Is spam: " << val << std::endl;
+    setup_perl(perl_int);
 
     string token = "";
-    int status = getToken("keys.cfg", token);
+    int status = get_token("keys.cfg", token);
     if (status) {
         std::cout << "Error in config file. Exitting..." << std::endl;
         return status;
@@ -103,7 +119,7 @@ int main(int argc, char **argv, char **env) {
     std::cout << "Telegram Token -> [" << token << "]" << std::endl;
 
     Bot bot(token);
-    setupBot(bot);
+    setup_bot(bot);
 
     signal(SIGINT, [](int sig) -> void {
         std::cout << "Got SIGINT exitting after finishing current poll." << std::endl;
@@ -117,6 +133,7 @@ int main(int argc, char **argv, char **env) {
         while (!sigintGot) {
             longPoll.start();
         }
+        destroy_perl(perl_int);
     } catch (exception& e) {
         std::cout << "Error: " << e.what() << std::endl;
     }
